last test cases

+++ src/tightdb/objc/test/query.m
@@ -65,15 +65,15 @@ TIGHTDB_TABLE_9(TestQueryAllTypes,
     STAssertEquals([query.IntCol sum], (int64_t)54,    @"IntCol sum");
     STAssertEquals([query.IntCol average] , 54.0,           @"IntCol avg");
 
-    STAssertEquals([query.FloatCol minimum] , 0.7f,         @"FloatCol min");
-    STAssertEquals([query.FloatCol maximum] , 0.7f,         @"FloatCol max");
-    STAssertEquals([query.FloatCol sum] , 0.7f, @"FloatCol sum");
-    STAssertEquals([query.FloatCol average] , (double)0.7f, @"FloatCol avg");
+    STAssertEquals([query.FloatCol minimum], 0.7f,         @"FloatCol min");
+    STAssertEquals([query.FloatCol maximum], 0.7f,         @"FloatCol max");
+    STAssertEquals([query.FloatCol sum], (double)0.7f, @"FloatCol sum");
+    STAssertEquals([query.FloatCol average], (double)0.7f, @"FloatCol avg");
 
-    STAssertEquals([query.DoubleCol minimum] , 0.8,         @"DoubleCol min");
-    STAssertEquals([query.DoubleCol maximum] , 0.8,         @"DoubleCol max");
-    STAssertEquals([query.DoubleCol sum] , 0.8,         @"DoubleCol sum");
-    STAssertEquals([query.DoubleCol average] , 0.8,         @"DoubleCol avg");
+    STAssertEquals([query.DoubleCol minimum], 0.8,         @"DoubleCol min");
+    STAssertEquals([query.DoubleCol maximum], 0.8,         @"DoubleCol max");
+    STAssertEquals([query.DoubleCol sum], 0.8,         @"DoubleCol sum");
+    STAssertEquals([query.DoubleCol average], 0.8,         @"DoubleCol avg");
 
     // Check that all column conditions return query objects of the
     // right type
